+++ tests/django/conftest.py
@@ -1,3 +1,7 @@
 import os
 
+import django
+
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.django_settings")
+
+django.setup()